Minor Changes
- Removed redundant files
- Reorganised folders and globals
- Added some exports
- Cleaned up navigation a bit
