--- lib/config/application.dart
@@ -1,5 +1,0 @@
-import 'package:fluro/fluro.dart';
-
-class Application {
-  static Router router;
-}
