+++ lib/redux/actions/action.dart
@@ -1,4 +1,5 @@
-abstract class Action {
+export 'ui_actions.dart';
 
+abstract class Action {
   String getDescription();
 }
