+++ lib/redux/reducers/app_reducer.dart
@@ -3,10 +3,11 @@ import 'package:community/redux/app_state.dart';
 import 'package:redux/redux.dart';
 
 final appReducer = combineReducers<AppState>([
-  TypedReducer<AppState,ChangeApplicationNameAction>(_changeApplicationNameReducer)
+  TypedReducer<AppState, ChangeApplicationNameAction>(
+      _changeApplicationNameReducer)
 ]);
 
-
-AppState _changeApplicationNameReducer(AppState state, ChangeApplicationNameAction action) {
+AppState _changeApplicationNameReducer(
+    AppState state, ChangeApplicationNameAction action) {
   return state.copyWith(appName: action.appName);
 }
