+++ lib/view/appComponent.dart
@@ -1,8 +1,7 @@
-import 'package:community/config/application.dart';
-import 'package:community/config/customTheme.dart';
-import 'package:community/config/globals.dart';
-import 'package:community/config/routes.dart';
 import 'package:community/redux/app_state.dart';
+import 'package:community/vendor//routes.dart';
+import 'package:community/vendor//themes.dart';
+import 'package:community/vendor/globals.dart';
 import 'package:dynamic_theme/dynamic_theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
@@ -29,8 +28,7 @@ class AppComponent extends StatefulWidget {
 class _AppComponentState extends State<AppComponent> {
   _AppComponentState() {
     // Configure Application Routes Navigator
-    Routes.configureRoutes(router);
-    Application.router = router;
+    Routes().defineRoutes(router);
   }
 
   @override
@@ -45,7 +43,7 @@ class _AppComponentState extends State<AppComponent> {
               debugShowCheckedModeBanner: false,
               title: widget.store.state.appName,
               theme: themeData,
-              onGenerateRoute: Application.router.generator,
+              onGenerateRoute: router.generator,
             ),
           );
         });
